feat(http_funcs): check the MIME type on uploaded files
The file payloads attached to POST requests are run through http.DetectContentType and then
subjected to an allowlist. There are a few new errors in the package to distinguish between some
different scenarious but ultimately, any upload that is not a ZIP file results in one of the
aforementioned errors.

part of resolving TT-1331