fix(api & http_funcs): move restarts to server handlers, and stop defer of cleanup
The overseer.Restart() calls were racing against cleanup operations, so this way the handlers will
call for a restart if and when necessary, and only after other operations (including temp file
cleanup) are complete. Also moved the cleanup calls outside of defer statements so that they would
run before the restarts, not during or after.

re TT-1594